feat: add toggle theme button and refactor logic to a hook
